+++ src/document.ts
@@ -5,16 +5,26 @@ export function getCurrentContext() {
   const workbenchConfig = vscode.workspace.getConfiguration('worldspider');
   const prefixLength = workbenchConfig.get('generation.prefixLength');
   const suffixLength = workbenchConfig.get('generation.suffixLength');
-  // const prefixLength = 100;
-  // const suffixLength = 100;
   let editor = vscode.window.activeTextEditor;
   if(!editor) {
     return {prefix: '', suffix: ''};
   }
   let context = editor.document.getText();
-  // get last prefixLength characters of prefix
   let currentContextPrefix = context.slice(0, editor.document.offsetAt(editor.selection.start)).slice(-<number>prefixLength);
-  // get first suffixLength characters of suffix
   let currentContextSuffix = context.slice(editor.document.offsetAt(editor.selection.end)).slice(0, <number>suffixLength);
   return {prefix: currentContextPrefix, suffix: currentContextSuffix};
+}
+
+export function getSelectionInfo() {
+  // returns currently selected text and absolute position in the document of the selection
+  let editor = vscode.window.activeTextEditor;
+  if(!editor) {
+    return {prefix: '', suffix: ''};
+  }
+  let selectedText = editor.document.getText(editor.selection);
+  let selectedStart = editor.selection.start;
+  let selectedEnd = editor.selection.end;
+  let selectedStartOffset = editor.document.offsetAt(selectedStart);
+  let selectedEndOffset = editor.document.offsetAt(selectedEnd);
+  return {selectedText, selectedStart, selectedEnd, selectedStartOffset, selectedEndOffset};
 }